+++ src/tightdb/column_basic.hpp
@@ -99,8 +99,7 @@ public:
 #ifdef TIGHTDB_DEBUG
     void Verify() const;
     void to_dot(std::ostream&, StringData title) const TIGHTDB_OVERRIDE;
-    void dump_node_structure(std::ostream&, int level) const TIGHTDB_OVERRIDE;
-    using ColumnBase::dump_node_structure;
+    void do_dump_node_structure(std::ostream&, int) const TIGHTDB_OVERRIDE;
 #endif
 
 protected: