+++ src/tightdb/index_string.hpp
@@ -27,37 +27,129 @@
 
 namespace tightdb {
 
+    // todo, add some type safety for these
+template <class T> inline StringData to_string(T value)
+{   
+    return StringData(reinterpret_cast<const char*>(&value), sizeof(T));
+}
+
+template <> inline StringData to_string(const char* value)
+{
+    return StringData(value);
+}
+
+template <> inline StringData to_string<StringData>(StringData input)
+{
+    return input;
+}
+
 typedef StringData (*StringGetter)(void*, std::size_t);
 
 class StringIndex: public Column {
 public:
     StringIndex(void* target_column, StringGetter get_func, Allocator&);
     StringIndex(ref_type, ArrayParent*, std::size_t ndx_in_parent, void* target_column,
-                StringGetter get_func, Allocator&);
+                StringGetter get_func, bool allow_duplicate_values, Allocator&);
     ~StringIndex() TIGHTDB_NOEXCEPT TIGHTDB_OVERRIDE {}
     void set_target(void* target_column, StringGetter get_func) TIGHTDB_NOEXCEPT;
 
     bool is_empty() const;
 
-    void insert(size_t row_ndx, StringData value, size_t num_rows, bool is_append);
-    void set(size_t row_ndx, StringData oldValue, StringData new_value);
-    void erase(size_t row_ndx, StringData value, bool is_last = false);
+    template <class T> void insert(size_t row_ndx, T value, size_t num_rows, bool is_append)
+    {
+        // If the new row is inserted after the last row in the table, we don't need
+        // to adjust any row indexes.
+        if (!is_append) {
+            for (size_t i = 0; i < num_rows; ++i) {
+                size_t row_ndx_2 = row_ndx + i;
+                adjust_row_indexes(row_ndx_2, 1); // Throws
+            }
+        }
+
+        for (size_t i = 0; i < num_rows; ++i) {
+            size_t row_ndx_2 = row_ndx + i;
+            size_t offset = 0; // First key from beginning of string
+            insert_with_offset(row_ndx_2, to_string(value), offset); // Throws
+        }
+    }
+
+    template <class T> void set(size_t row_ndx, T old_value, T new_value)
+    {
+        StringData old_value2 = to_string(old_value);
+        StringData new_value2 = to_string(new_value);
+        // Note that insert_with_offset() throws UniqueConstraintViolation.
+
+        if (TIGHTDB_LIKELY(new_value2 != old_value2)) {
+            size_t offset = 0; // First key from beginning of string
+            insert_with_offset(row_ndx, new_value2, offset); // Throws
+
+            bool is_last = true; // To avoid updating refs
+            erase(row_ndx, old_value2, is_last); // Throws
+        }
+    }
+
+    template <class T> size_t find_first(T value) const
+    {
+        // Use direct access method
+        return m_array->IndexStringFindFirst(to_string(value), m_target_column, m_get_func);
+    }
+
+    template <class T> void find_all(Column& result, T value) const
+    {
+        // Use direct access method
+        return m_array->IndexStringFindAll(result, value, m_target_column, m_get_func);
+    }
+
+    template <class T> FindRes find_all(T value, size_t& ref) const
+    {
+        // Use direct access method
+        return m_array->IndexStringFindAllNoCopy(to_string(value), ref, m_target_column, m_get_func);
+    }
+
+    template <class T> size_t count(T value) const
+    {
+        // Use direct access method
+        return m_array->IndexStringCount(to_string(value), m_target_column, m_get_func);
+    }
+
+    template <class T> void erase(size_t row_ndx, T value, bool is_last)
+    {
+        DoDelete(row_ndx, to_string(value), 0);
+
+        // Collapse top nodes with single item
+        while (!root_is_leaf()) {
+            TIGHTDB_ASSERT(m_array->size() > 1); // node cannot be empty
+            if (m_array->size() > 2)
+                break;
+
+            ref_type ref = m_array->get_as_ref(1);
+            m_array->set(1, 1); // avoid destruction of the extracted ref
+            m_array->destroy_deep();
+            m_array->init_from_ref(ref);
+            m_array->update_parent();
+        }
+
+        // If it is last item in column, we don't have to update refs
+        if (!is_last)
+            adjust_row_indexes(row_ndx, -1);
+    }
+
+    template <class T> void update_ref(T value, size_t old_row_ndx, size_t new_row_ndx)
+    {
+        do_update_ref(to_string(value), old_row_ndx, new_row_ndx, 0);
+    }
+
     void clear() TIGHTDB_OVERRIDE;
+    void distinct(Column& result) const;
+    bool has_duplicate_values() const TIGHTDB_NOEXCEPT;
 
-    using Column::insert;
-    using Column::erase;
-
-    size_t count(StringData value) const;
-    size_t find_first(StringData value) const;
-    void   find_all(Column& result, StringData value) const;
-    void   distinct(Column& result) const;
-    FindRes find_all(StringData value, size_t& ref) const;
-
-    void update_ref(StringData value, size_t old_row_ndx, size_t new_row_ndx);
+    /// By default, duplicate values are allowed.
+    void set_allow_duplicate_values(bool) TIGHTDB_NOEXCEPT;
 
 #ifdef TIGHTDB_DEBUG
     void Verify() const TIGHTDB_OVERRIDE;
     void verify_entries(const AdaptiveStringColumn& column) const;
+    void do_dump_node_structure(std::ostream&, int) const TIGHTDB_OVERRIDE;
     void to_dot() const { to_dot(std::cerr); }
     void to_dot(std::ostream&, StringData title = StringData()) const;
 #endif
@@ -67,15 +159,25 @@ public:
     static key_type create_key(StringData) TIGHTDB_NOEXCEPT;
 
 private:
+    void* m_target_column;
+    StringGetter m_get_func;
+    bool m_deny_duplicate_values;
+
+    using Column::insert;
+    using Column::erase;
+
     struct inner_node_tag {};
     StringIndex(inner_node_tag, Allocator&);
 
     static Array* create_node(Allocator&, bool is_leaf);
 
-    void InsertWithOffset(size_t row_ndx, size_t offset, StringData value);
+    void insert_with_offset(size_t row_ndx, StringData value, size_t offset);
     void InsertRowList(size_t ref, size_t offset, StringData value);
     key_type GetLastKey() const;
-    void UpdateRefs(size_t pos, int diff);
+
+    /// Add small signed \a diff to all elements that are greater than, or equal
+    /// to \a min_row_ndx.
+    void adjust_row_indexes(size_t min_row_ndx, int diff);
 
     struct NodeChange {
         size_t ref1;
@@ -97,13 +199,10 @@ private:
 
     StringData get(size_t ndx) {return (*m_get_func)(m_target_column, ndx);}
 
-    // Member variables
-    void* m_target_column;
-    StringGetter m_get_func;
-
     void NodeAddKey(ref_type ref);
 
 #ifdef TIGHTDB_DEBUG
+    static void dump_node_structure(const Array& node, std::ostream&, int level);
     void to_dot_2(std::ostream&, StringData title = StringData()) const;
     static void array_to_dot(std::ostream&, const Array&);
     static void keys_to_dot(std::ostream&, const Array&, StringData title = StringData());
@@ -118,27 +217,35 @@ private:
 inline StringIndex::StringIndex(void* target_column, StringGetter get_func, Allocator& alloc):
     Column(create_node(alloc, true)), // Throws
     m_target_column(target_column),
-    m_get_func(get_func)
-{
-}
-
-inline StringIndex::StringIndex(inner_node_tag, Allocator& alloc):
-    Column(create_node(alloc, false)), // Throws
-    m_target_column(0),
-    m_get_func(0)
+    m_get_func(get_func),
+    m_deny_duplicate_values(false)
 {
 }
 
 inline StringIndex::StringIndex(ref_type ref, ArrayParent* parent, std::size_t ndx_in_parent,
-                                void* target_column, StringGetter get_func, Allocator& alloc):
+                                void* target_column, StringGetter get_func,
+                                bool deny_duplicate_values, Allocator& alloc):
     Column(alloc, ref),
     m_target_column(target_column),
-    m_get_func(get_func)
+    m_get_func(get_func),
+    m_deny_duplicate_values(deny_duplicate_values)
 {
     TIGHTDB_ASSERT(Array::get_context_flag_from_header(alloc.translate(ref)));
     set_parent(parent, ndx_in_parent);
 }
 
+inline StringIndex::StringIndex(inner_node_tag, Allocator& alloc):
+    Column(create_node(alloc, false)), // Throws
+    m_target_column(0),
+    m_get_func(0),
+    m_deny_duplicate_values(false)
+{
+}
+
+inline void StringIndex::set_allow_duplicate_values(bool allow) TIGHTDB_NOEXCEPT
+{
+    m_deny_duplicate_values = !allow;
+}
 
 inline StringIndex::key_type StringIndex::create_key(StringData str) TIGHTDB_NOEXCEPT
 {